fix(l10n): clean up currency name localizations (#449)
* fix(l10n): remove unused currency l10n files and clean up code

Deleted several unused currency localization classes for various languages
to streamline the localization process and reduce code clutter.

* fix(docs): update translation counts for quechua and northern sami locales

Adjusted the number of translations documented in the comments for both
QuCurrenciesL10N and SeCurrenciesL10N classes to reflect the actual
translations provided.

* fix(l10n): correct capitalization of currency names in belarusian locale

Updated the currency names in the Belarusian localization to ensure proper capitalization.

* fix(l10n): correct capitalization of currency names in faroese and canadian french

* fix: update currency names for fr, hr, nn and no locales for consistency

Updated the currency names in the nn_currencies_l10n.data.dart and
no_currencies_l10n.data.dart files to ensure proper capitalization
and consistency in the translations.

* fix(l10n): update currency translations for to and uz cyrl locales

- corrected the number of translations for the `to` locale
- removed unused currency entries from the `uz_cyrl` locale

* fix(test): update expected currency map length in tests

Adjusted the expected length of the currency map in the constructor and custom translations tests to reflect recent changes in the data.

* fix(l10n): update translation for russian currency in croatian locale

changed "Ruska rublja" to "Ruski rubalj" for consistency.